streamlit_app/requirements.txt
streamlit
databricks-sql-connector
pandas # Often useful for data manipulation and display
plotly # If you plan to use Plotly for charts